test: share cli diag assertion helpers
- extract diag mode parsing/assertion into tests/common module
- reuse shared helper in topology_fallback integration tests
- record test-infra progress in backlog
